+++ libraries/dynamic-ui-react/src/components/banking/index.ts
@@ -1,4 +1,3 @@
-export { default as MCollapseIconText } from './MCollapseIconText';
 export { default as MPermissionGroup } from './MPermissionGroup';
 export { default as DPermissionItem } from './DPermissionItem';
 export { default as MSummaryCard } from './MSummaryCard';